# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Vaca EFA API
  version: 0.1.0
  description: API del sistema de control de vacaciones, saldos derivados y trazabilidad laboral.
servers:
  - url: http://localhost:3000
    description: Desarrollo local
security:
  - cookieAuth: []
paths:
...
  /api/v1/holidays:
    get:
      summary: Consultar festivos por año con una sesión de lectura
      parameters: [{name: year, in: query, schema: {type: integer, minimum: 2000}}]
      responses: {'200': {description: Festivos configurados}}
  /api/v1/admin/holidays:
    get:
      summary: Listar festivos administrativos
      parameters: [{name: year, in: query, schema: {type: integer, minimum: 2000}}]
      responses: {'200': {description: Festivos configurados}}
    post:
      summary: Crear festivo
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/HolidayInput'}}}
      responses: {'201': {description: Festivo creado}}
  /api/v1/admin/holidays/{id}:
    get:
      summary: Obtener festivo
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: Festivo}, '404': {description: No encontrado}}
    patch:
      summary: Editar festivo
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/HolidayPatch'}}}
      responses: {'200': {description: Festivo actualizado}}
    delete:
      summary: Desactivar festivo conservando auditoría
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: Festivo desactivado}}
  /api/v1/admin/users:
    get:
      summary: Listar usuarios sin credenciales sensibles
      responses: {'200': {description: Usuarios administrativos}}
    post:
      summary: Crear usuario administrativo
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/UserCreate'}}}
      responses: {'201': {description: Usuario creado}, '409': {description: Usuario duplicado}}
  /api/v1/admin/users/{id}:
    patch:
      summary: Editar, activar o inactivar usuario sin eliminarlo
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/UserPatch'}}}
      responses: {'200': {description: Usuario actualizado}, '409': {description: Usuario duplicado}, '422': {description: Regla de seguridad de administradores}}
  /api/v1/admin/settings/{key}:
    get:
      summary: Consultar un ajuste del sistema
      description: "Claves admitidas: GERENTE (firma del PDF anual) y VACATION_CLOSURE_FROM_DATE (fecha de corte de los cierres históricos, formato YYYY-MM-DD, por defecto 2025-01-01)."
      parameters:
        - name: key
          in: path
          required: true
          schema: {type: string, enum: [GERENTE, VACATION_CLOSURE_FROM_DATE]}
      responses:
        '200': {description: Ajuste persistido o vacío cuando no existe}
        '404': {description: Clave no admitida}
    patch:
      summary: Guardar un ajuste del sistema
      description: VACATION_CLOSURE_FROM_DATE exige una fecha válida YYYY-MM-DD. El cambio queda auditado como SYSTEM_SETTING_UPDATED.
      parameters:
        - name: key
          in: path
          required: true
          schema: {type: string, enum: [GERENTE, VACATION_CLOSURE_FROM_DATE]}
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, required: [value], properties: {value: {type: string}}}
      responses:
        '200': {description: Ajuste guardado}
        '400': {$ref: '#/components/responses/ValidationError'}
        '404': {description: Clave no admitida}
  /api/v1/alerts:
    get:
      summary: Consultar alertas persistidas
      parameters:
        - {name: employmentId, in: query, schema: {type: string}}
        - {name: active, in: query, schema: {type: boolean}}
      responses: {'200': {description: Alertas}}
...
  /api/v1/admin/retired-employments/close-pending:
    post:
      summary: Cerrar pendientes de vacaciones de vínculos retirados
      description: Operación ADMIN idempotente. Lee períodos por lote, cierra el ciclo a la fecha de retiro y conserva auditoría.
      responses:
        '200': {description: Resultado de regularización}
        '403': {description: Se requiere permiso administrativo}
...
  /api/v1/audit/page:
    get:
      summary: Auditoría paginada
      parameters:
        - in: query
          name: page
          schema: {type: integer, minimum: 1, default: 1}
        - in: query
          name: pageSize
          schema: {type: integer, minimum: 1, maximum: 200, default: 20}
        - in: query
          name: actorId
          schema: {type: string}
        - in: query
          name: action
          schema: {type: string}
        - in: query
          name: entityType
          schema: {type: string}
        - in: query
          name: entityId
          schema: {type: string}
        - in: query
          name: fromDate
          schema: {type: string, format: date}
        - in: query
          name: toDate
          schema: {type: string, format: date}
      responses:
        '200':
          description: Página de eventos de auditoría
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PageDto_AuditEvent'}
        '401': {$ref: '#/components/responses/Unauthorized'}
...
  /api/v1/holidays/page:
    get:
      summary: Festivos paginados (vista autenticada)
      parameters:
        - {in: query, name: page, schema: {type: integer, minimum: 1, default: 1}}
        - {in: query, name: pageSize, schema: {type: integer, minimum: 1, maximum: 200, default: 20}}
        - {in: query, name: year, schema: {type: integer, minimum: 2000, maximum: 2100}}
        - {in: query, name: active, schema: {type: boolean}}
        - {in: query, name: search, schema: {type: string}}
      responses:
        '200':
          description: Página de festivos
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PageDto_Holiday'}
        '401': {$ref: '#/components/responses/Unauthorized'}
...
  /api/v1/admin/users/page:
    get:
      summary: Usuarios paginados
      parameters:
        - {in: query, name: page, schema: {type: integer, minimum: 1, default: 1}}
        - {in: query, name: pageSize, schema: {type: integer, minimum: 1, maximum: 200, default: 20}}
        - {in: query, name: search, schema: {type: string}}
        - {in: query, name: active, schema: {type: boolean}}
        - in: query
          name: role
          schema:
            type: string
            enum: [ADMIN, HR, VIEWER, READ_ONLY]
      responses:
        '200':
          description: Página de usuarios
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PageDto_UserSummary'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/admin/catalogs/{type}/page:
    get:
      summary: Valores de catálogo paginados
      parameters:
        - {in: path, name: type, required: true, schema: {type: string}}
        - {in: query, name: page, schema: {type: integer, minimum: 1, default: 1}}
        - {in: query, name: pageSize, schema: {type: integer, minimum: 1, maximum: 200, default: 20}}
        - {in: query, name: active, schema: {type: boolean}}
        - {in: query, name: search, schema: {type: string}}
      responses:
        '200':
          description: Página de catálogo
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PageDto_CatalogItem'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/alerts/page:
    get:
      summary: Alertas paginadas
      parameters:
        - {in: query, name: page, schema: {type: integer, minimum: 1, default: 1}}
        - {in: query, name: pageSize, schema: {type: integer, minimum: 1, maximum: 200, default: 20}}
        - {in: query, name: employmentId, schema: {type: string}}
        - {in: query, name: severity, schema: {type: string, enum: [INFO, WARNING, CRITICAL]}}
        - {in: query, name: type, schema: {type: string, enum: [UPCOMING_ACCRUAL, OVERDUE_PERIOD, UPCOMING_VACATION, PENDING_AND_UPCOMING]}}
        - {in: query, name: active, schema: {type: boolean}}
      responses:
        '200':
          description: Página de alertas
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PageDto_VacationAlert'}
  /api/v1/admin/scheduler-runs/page:
    get:
      summary: Ejecuciones del scheduler paginadas
      parameters:
        - {in: query, name: page, schema: {type: integer, minimum: 1, default: 1}}
        - {in: query, name: pageSize, schema: {type: integer, minimum: 1, maximum: 200, default: 20}}
        - {in: query, name: jobName, schema: {type: string, enum: [VACATION_ACCRUAL]}}
        - {in: query, name: status, schema: {type: string, enum: [RUNNING, COMPLETED, FAILED]}}
        - {in: query, name: fromDate, schema: {type: string, format: date}}
        - {in: query, name: toDate, schema: {type: string, format: date}}
      responses:
        '200':
          description: Página de ejecuciones
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PageDto_SchedulerRun'}
  /api/v1/admin/import-batches/page:
    get:
      summary: Lotes de importación de empleados paginados
      parameters:
        - {in: query, name: page, schema: {type: integer, minimum: 1, default: 1}}
        - {in: query, name: pageSize, schema: {type: integer, minimum: 1, maximum: 200, default: 20}}
        - in: query
          name: status
          schema: {type: string, enum: [PROCESSING, COMPLETED, COMPLETED_WITH_ERRORS, FAILED]}
        - {in: query, name: fileName, schema: {type: string}}
        - {in: query, name: fromDate, schema: {type: string, format: date}}
        - {in: query, name: toDate, schema: {type: string, format: date}}
      responses:
        '200':
          description: Página de lotes
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PageDto_ImportBatchSummary'}
  /api/v1/admin/retired-employments/reconciliation/page:
    get:
      summary: Página de conciliación de empleados retirados
      parameters:
        - {in: query, name: page, schema: {type: integer, minimum: 1, default: 1}}
        - {in: query, name: pageSize, schema: {type: integer, minimum: 1, maximum: 200, default: 20}}
        - {in: query, name: search, schema: {type: string}}
        - in: query
          name: state
          schema: {type: string, enum: [ENJOYED, LIQUIDATED, PENDING_LIQUIDATION]}
        - {in: query, name: asOf, schema: {type: string, format: date}}
      responses:
        '200':
          description: Página de conciliación
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RetiredReconciliationItemPage'}
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: vaca_access
  responses:
    Unauthorized:
      description: Sesión ausente, inválida o expirada
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
...
    ValidationError:
      description: Cuerpo, query o límite de entrada inválido
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
...
  parameters:
    Id: {name: id, in: path, required: true, schema: {type: string}}
...
  schemas:
...
    UserCreate:
      type: object
      required: [username, password, role]
      properties:
        username: {type: string, minLength: 3}
        displayName: {type: string, minLength: 2}
        jobTitle: {type: string, minLength: 2}
        password: {type: string, minLength: 8, format: password}
        role: {type: string, enum: [ADMIN, HR, VIEWER, READ_ONLY]}
    UserPatch:
      type: object
      properties:
        username: {type: string, minLength: 3}
        displayName: {type: string, minLength: 2}
        jobTitle: {type: string, minLength: 2}
        password: {type: string, minLength: 8, format: password}
        role: {type: string, enum: [ADMIN, HR, VIEWER, READ_ONLY]}
        active: {type: boolean}
    HolidayInput:
      type: object
      required: [date, name]
      properties:
        date: {type: string, format: date}
        name: {type: string}
        country: {type: string, default: CO}
        active: {type: boolean, default: true}
    HolidayPatch:
      type: object
      properties:
        date: {type: string, format: date}
        name: {type: string}
        country: {type: string}
        active: {type: boolean}
      description: Todos los campos son opcionales en una actualización.
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, BUSINESS_RULE_VIOLATION, INTERNAL_ERROR]
        message: {type: string}
        requestId: {type: string, description: Correlación de la petición; puede faltar en respuestas de compatibilidad directas}
        details:
          type: object
          additionalProperties: true
...
    PageDto:
      type: object
      required: [items, page, pageSize, total, hasNext]
      properties:
        items: {type: array, items: {}}
        page: {type: integer, minimum: 1}
        pageSize: {type: integer, minimum: 1}
        total: {type: integer, minimum: 0}
        hasNext: {type: boolean}
...
    PageDto_AuditEvent:
      allOf:
        - {$ref: '#/components/schemas/PageDto'}
        - type: object
          properties:
            items: {type: array, items: {$ref: '#/components/schemas/AuditEvent'}}
    PageDto_Holiday:
      allOf:
        - {$ref: '#/components/schemas/PageDto'}
        - type: object
          properties:
            items: {type: array, items: {$ref: '#/components/schemas/Holiday'}}
    PageDto_UserSummary:
      allOf:
        - {$ref: '#/components/schemas/PageDto'}
        - type: object
          properties:
            items: {type: array, items: {$ref: '#/components/schemas/UserSummary'}}
    PageDto_CatalogItem:
      allOf:
        - {$ref: '#/components/schemas/PageDto'}
        - type: object
          properties:
            items: {type: array, items: {$ref: '#/components/schemas/CatalogItem'}}
    PageDto_VacationAlert:
      allOf:
        - {$ref: '#/components/schemas/PageDto'}
        - type: object
          properties:
            items: {type: array, items: {$ref: '#/components/schemas/VacationAlert'}}
    PageDto_SchedulerRun:
      allOf:
        - {$ref: '#/components/schemas/PageDto'}
        - type: object
          properties:
            items: {type: array, items: {$ref: '#/components/schemas/SchedulerRun'}}
    PageDto_ImportBatchSummary:
      allOf:
        - {$ref: '#/components/schemas/PageDto'}
        - type: object
          properties:
            items: {type: array, items: {$ref: '#/components/schemas/ImportBatchSummary'}}
    RetiredReconciliationItemPage:
      allOf:
        - {$ref: '#/components/schemas/PageDto'}
        - type: object
          properties:
            items: {type: array, items: {$ref: '#/components/schemas/RetiredReconciliationItem'}}
    AuditEvent:
      type: object
      required: [id, actorId, action, entityType, entityId, createdAt]
      properties:
        id: {type: string}
        actorId: {type: string}
        action: {type: string}
        entityType: {type: string}
        entityId: {type: string}
        metadata: {type: object, additionalProperties: true}
        createdAt: {type: string, format: date-time}
...